+++ vsd/src/downloader/parse.rs
@@ -1,6 +1,7 @@
 use super::fetch::Metadata;
 use crate::{
-    playlist::{AutomationOptions, MasterPlaylist, MediaPlaylist, PlaylistType},
+    automation::{Prompter, SelectOptions},
+    playlist::{MasterPlaylist, MediaPlaylist, PlaylistType},
     utils,
 };
 use anyhow::{Result, anyhow, bail};
@@ -121,11 +122,12 @@ pub fn parse_all_streams(
 }
 
 pub fn parse_selected_streams(
-    auto_opts: &AutomationOptions,
     base_url: Option<Url>,
     client: &Client,
     meta: &Metadata,
+    prompter: &Prompter,
     query: &HashMap<String, String>,
+    mut select_opts: SelectOptions,
 ) -> Result<Vec<MediaPlaylist>> {
     match meta.pl_type {
         Some(PlaylistType::Dash) => {
@@ -133,7 +135,7 @@ pub fn parse_selected_streams(
                 .map_err(|_| anyhow!("couldn't parse response ({}) as dash playlist.", meta.url))?;
             let mut streams = crate::dash::parse_as_master(&mpd, meta.url.as_ref())
                 .sort_streams()
-                .select_streams(auto_opts)?;
+                .select_streams(prompter, &mut select_opts)?;
 
             for stream in &mut streams {
                 crate::dash::push_segments(
@@ -155,7 +157,7 @@ pub fn parse_selected_streams(
             Ok(m3u8_rs::Playlist::MasterPlaylist(m3u8)) => {
                 let mut streams = crate::hls::parse_as_master(&m3u8, meta.url.as_str())
                     .sort_streams()
-                    .select_streams(auto_opts)?;
+                    .select_streams(prompter, &mut select_opts)?;
 
                 for stream in &mut streams {
                     stream.uri = base_url